Pewpewpew V1.0
Need to improve the patch's ability to identify the note onset trigged by the electric guitar. In this version there are two knobs - threshold and gate - that allows the user to adjust sensitivity of note onset detection. In the most ideal scenario, the user can store presets with a fixed threshold and gate designed for each guitar or tone setup that the user will be using in a set.